+++ foxes/input/states/weibull_field.py
@@ -276,34 +276,6 @@ class WeibullField(FieldData):
         data0 = data
 
         return None, x, y, h, data
-    
-    def load_data(self, algo, verbosity=0):
-        """
-        Load and/or create all model data that is subject to chunking.
-
-        Such data should not be stored under self, for memory reasons. The
-        data returned here will automatically be chunked and then provided
-        as part of the mdata object during calculations.
-
-        Parameters
-        ----------
-        algo: foxes.core.Algorithm
-            The calculation algorithm
-        verbosity: int
-            The verbosity level, 0 = silent
-
-        Returns
-        -------
-        idata: dict
-            The dict has exactly two entries: `data_vars`,
-            a dict with entries `name_str -> (dim_tuple, data_ndarray)`;
-            and `coords`, a dict with entries `dim_name_str -> dim_array`
-
-        """
-        idata = super().load_data(algo, verbosity)
-        if FC.STATE in idata["coords"]:
-            del idata["coords"][FC.STATE]
-        return idata
 
     def output_point_vars(self, algo):
         """